Remove -Xshare:on from Java flags
Xshare:on breaks the plugin on Raspberry Pi:
https://github.com/ev3dev-lang-java/gradle-plugin/issues/2

Apparently the CDS cache is sometimes not generated properly
on Raspbian. This means that we either have to generate
the cache in our tooling, or that we have to disable -Xshare:on.

I think the best solution is to use default configuration.
https://docs.oracle.com/javase/9/vm/class-data-sharing.htm :

-Xshare:auto
    The default; enable class data sharing whenever possible.

This means that on EV3, it should pick up the cache generated
by JRI package postinst or by the tarball installation script.
On systems where the CDS cache is not available, this ensures
that the execution proceeds, albeit it will have a bit slower
startup.

We can combine this with change with the introduction of
-Xshare:dump for Raspberry's in installer.sh. After that
is realized, this commit ensures that old setups will work
as intended.

+ bump to 1.6.5